iccsi/requirements.txt
Django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow>=10.0.0
stripe==7.8.0
djangorestframework-simplejwt==5.3.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
